Adding new mail folder

+++ src/main/java/org/luwrain/settings/mail/folders/Element.java
@@ -7,10 +7,10 @@ import org.luwrain.cpanel.*;
 public class Element implements org.luwrain.cpanel.Element
 {
 public final org.luwrain.cpanel.Element parent;
-public final long id;
+public final int id;
 public final String title;
 
-Element(org.luwrain.cpanel.Element parent, long id, String title)
+Element(org.luwrain.cpanel.Element parent, int id, String title)
     {
 	NullCheck.notNull(parent, "parent");
 	NullCheck.notNull(title, "title");
